+++ examples/Program.cs
@@ -125,6 +125,10 @@ class Program
         await UpsertsWithReplacingMergeTree.Run();
         WaitForUser(isInteractive);
 
+        Console.WriteLine($"\n\nRunning: {nameof(SchemaOptimization)}");
+        await SchemaOptimization.Run();
+        WaitForUser(isInteractive);
+
         // Selecting Data
         Console.WriteLine("\n\n" + new string('=', 70));
         Console.WriteLine("SELECTING DATA");